fix: module not found errors

+++ fileUtils.js
@@ -1,14 +1,13 @@
-import { rename, unlink, writeFile } from 'node:fs/promises';
-import { createReadStream, createWriteStream } from 'node:fs';
+import { cp, rename, unlink, writeFile } from 'node:fs/promises';
+import { createReadStream } from 'node:fs';
 import { resolve, join } from 'node:path';
-import { pipeline } from 'node:stream';
-import { printCurrentDirectory, printOperationFailed } from './consoleUtils';
+import { printCurrentDirectory, printOperationFailed } from './consoleUtils.js';
 
 export function readFile(currentDir, path) {
     const filePath = resolve(currentDir, path);
     const stream = createReadStream(filePath, { encoding: 'utf-8' });
     stream.on('data', chunk => console.log(chunk));
-    stream.on('end', () => printCurrentDirectory());
+    stream.on('end', () => printCurrentDirectory(currentDir));
     stream.on('error', () => printOperationFailed());
 }
 
@@ -16,7 +15,7 @@ export async function createFile(currentDir, name) {
     const filePath = join(currentDir, name);
     try{
         await writeFile(filePath, '');
-    }catch(e){
+    }catch{
         printOperationFailed();
     }
 }
@@ -26,33 +25,32 @@ export async function renameFile(currentDir, oldName, newName) {
     const newPath = resolve(currentDir, newName);
     try{
         await rename(oldPath, newPath);
-    }catch(e){
+    }catch{
         printOperationFailed();
     }
 }
 
-export function copyFile(currentDir, source, destination) {
+export async function copyFile(currentDir, source, destination, toDelete = false) {
     const sourcePath = resolve(currentDir, source);
-    const destPath = resolve(currentDir, destination, source);
+    const destPath = resolve(currentDir, destination);
 
-    const readStream = createReadStream(sourcePath);
-    const writeStream = createWriteStream(destPath);
-
-    pipeline(readStream, writeStream, (err) => {
-        if (err) printOperationFailed();
-    });
+    try{
+        await cp(sourcePath, destPath);
+        if (toDelete) await deleteFile(currentDir, source);
+    }catch{
+        printOperationFailed();
+    }
 }
 
 export async function moveFile(currentDir, source, destination) {
-    copyFile(currentDir, source, destination);
-    await deleteFile(currentDir, source);
+    await copyFile(currentDir, source, destination, true);
 }
 
 export async function deleteFile(currentDir, path) {
     const filePath = resolve(currentDir, path);
     try{
         await unlink(filePath);
-    }catch(e){
+    }catch{
         printOperationFailed();
     }
 }
